+++ CHANGES.md
@@ -74,6 +74,16 @@
   incorrect results on concrete values.
   ([#2044](https://github.com/GaloisInc/cryptol/issues/2044))
 
+## API changes
+
+* Add `isValidIdent` to `Cryptol.Parser.LexerUtils`, which checks if a name is
+  a valid Cryptol identifier.
+  ([#2036](https://github.com/GaloisInc/cryptol/issues/2036))
+
+* Add `pIsNeq` to `Cryptol.TypeCheck.Type`, which recognizes if a Cryptol
+  constraint is headed by a not-equal (`!=`) operator.
+  ([#2038](https://github.com/GaloisInc/cryptol/issues/2038))
+
 # 3.5.0 -- 2026-01-27
 
 ## Administrative changes